Issue #662: Fixed CMR parameters
